backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
google-generativeai>=0.8.0
aiosmtplib>=3.0.0
bleach>=6.0.0
